enhance styleguide to use addEventListener insetad of addListener
