+++ src/index.js
@@ -163,3 +163,4 @@ export {default as DialogBody} from './components/dialog/DialogBody';
 export type {DialogBodyPropsType} from './components/dialog/DialogBody';
 export {default as DialogCloseButton} from './components/dialog/DialogCloseButton';
 export type {DialogCloseButtonPropsType} from './components/dialog/DialogCloseButton';
+export {default as useReducedMotion} from './components/utils/useReducedMotion';
